Implemented feature #414 (EPIC loader truncating component code/native_code to 20 chars. Changing this to truncate to 70 chars).
In addition, when loading lab result data, the "native_code" is derived by concatenating "cpt" + "--" + "component".  Previously, if one of either cpt or component was missing, the "--" was not included in the native_code.  I've changed the loader to always include the "--" so that it is easier to distinguish cpt from component in the native_code.

Former-commit-id: 114d1b7610342f2acb88de50630702d2bdbca9d3