Removed references to "fasting-glucose-status" which is no longer used.
Removed "lx:ogtt75-fasting-urine:positive" from the set of labs referenced in the "ogtt75_intra_thresh_q" query.

Removed "lx:ogtt100-fasting-urine:positive" from the set of labs referenced in the "ogtt100_threshold_q" query.

Removed "ogtt75-fasting-urine" from the ogtt75_labname sequence.

Changed the test for nutrition referral to match the spec.  We now search the site name for "Nutr" (not "Nutrition").

Changed the calculations for the obfastingglucose_positive column to only consider lab results returned by the "obfastingglucose_q" query for which the test result reference high is between 90 and 99.  This column should not pay attention to test results associated with the "glucosefasting_q" or "ogttfasting_high_q" queries.

Changed the calculation for the postpartum_random GDM report column (now named postpartum_random_gte200_twice) to only be set to 1 if there are two or more random glucose tests above the threshold.

Renamed the following columns for the GDM report:
  prior_gdm_icd9_this_preg -> prior_gdm_icd9
  prior_gdm_icd9_this_preg_date -> prior_gdm_icd9_date
  postpartum_ogtt75_order -> postpartum120_ogtt75_order
  postpartum_ogtt75_any_result -> postpartum120_ogtt75_any_result
  postpartum_random -> postpartum_random_gte200_twice
  postpartum_random_date1 -> postpartum_random_gRemoved references to "fasting-glucose-status" which is no longer used.

Removed "lx:ogtt75-fasting-urine:positive" from the set of labs referenced in the "ogtt75_intra_thresh_q" query.

Removed "lx:ogtt100-fasting-urine:positive" from the set of labs referenced in the "ogtt100_threshold_q" query.

Removed "ogtt75-fasting-urine" from the ogtt75_labname sequence.

Changed the test for nutrition referral to match the spec.  We now search the site name for "Nutr" (not "Nutrition").

Changed the calculations for the obfastingglucose_positive column to only consider lab results returned by the "obfastingglucose_q" query for which the test result reference high is between 90 and 99.  This column should not pay attention to test results associated with the "glucosefasting_q" or "ogttfasting_high_q" queries.

Changed the calculation for the postpartum_random GDM report column (now named postpartum_random_gte200_twice) to only be set to 1 if there are two or more random glucose tests above the threshold.

Renamed the following columns for the GDM report:
  prior_gdm_icd9_this_preg -> prior_gdm_icd9
  prior_gdm_icd9_this_preg_date -> prior_gdm_icd9_date
  postpartum_ogtt75_order -> postpartum120_ogtt75_order
  postpartum_ogtt75_any_result -> postpartum120_ogtt75_any_result
  postpartum_random -> postpartum_random_gte200_twice
  postpartum_random_date1 -> postpartum_random_gte200_date1
  postpartum_random_date2 -> postpartum_random_gte200_date2
  postpartum_a1c -> postpartum_a1c_6_5
  postpartum_a1c_date -> postpartum_a1c_6_5_date

Added the following new columns to the GDM report:
  postpartum_ogtt75_any_result_date
  postpartum_fasting_any_result
  postpartum_fasting_gte100_lte125
  postpartum_fasting_gte100_lte125_max_val
  postpartum_fasting_gte100_lte125_date
  postpartum_fasting_gte126_max_val
  postpartum_a1c_any_result
  postpartum_a1c_gte5_7_lte6_4
  postpartum_a1c_gte5_7_lte6_4_max_val
  postpartum_a1c_gte5_7_lte6_4_date
  postpartum_a1c_6_5_max_val
  yearplus_postpartum_a1c_max_val
  any_a1c_gte5_7_lte6_4
  any_a1c_gte5_7_lte6_4_max_val
  any_a1c_gte5_7_lte6_4_date
  any_a1c_6_5_max_val

Changed the postpartum_ogtt75_* columns to use 0-256 for the postpartum period.  This affected the following columns in the GDM report:
  postpartum_ogtt75_order
  postpartum_ogtt75_any_result
  postpartum_ogtt75_fasting_max_val
  postpartum_ogtt75_2hr_max_val
  postpartum_ogtt75_ifg_range_met
  postpartum_ogtt75_igt_range_met
  postpartum_ogtt75_dm_range_met

Deleted the following columns from the GDM report:
  any_a1c_5_7
  any_a1c_5_7_date
  te200_date1
  postpartum_random_date2 -> postpartum_random_gte200_date2
  postpartum_a1c -> postpartum_a1c_6_5
  postpartum_a1c_date -> postpartum_a1c_6_5_date

Added the following new columns to the GDM report:
  postpartum_ogtt75_any_result_date
  postpartum_fasting_any_result
  postpartum_fasting_gte100_lte125
  postpartum_fasting_gte100_lte125_max_val
  postpartum_fasting_gte100_lte125_date
  postpartum_fasting_gte126_max_val
  postpartum_a1c_any_result
  postpartum_a1c_gte5_7_lte6_4
  postpartum_a1c_gte5_7_lte6_4_max_val
  postpartum_a1c_gte5_7_lte6_4_date
  postpartum_a1c_6_5_max_val
  yearplus_postpartum_a1c_max_val
  any_a1c_gte5_7_lte6_4
  any_a1c_gte5_7_lte6_4_max_val
  any_a1c_gte5_7_lte6_4_date
  any_a1c_6_5_max_val

Changed the postpartum_ogtt75_* columns to use 0-256 for the postpartum period.  This affected the following columns in the GDM report:
  postpartum_ogtt75_order
  postpartum_ogtt75_any_result
  postpartum_ogtt75_fasting_max_val
  postpartum_ogtt75_2hr_max_val
  postpartum_ogtt75_ifg_range_met
  postpartum_ogtt75_igt_range_met
  postpartum_ogtt75_dm_range_met

Deleted the following columns from the GDM report:
  any_a1c_5_7
  any_a1c_5_7_date


Former-commit-id: 36b488487362ededd01aedd3c2724e5db6652cb9